rename Add to AddVideo
